requirements.txt
decorator==4.0.2
Django==1.8.4
django-bootstrap3==6.2.2
django-countries==3.3
django-endless-pagination==2.0
django-extensions==1.5.6
django-money==0.7.2
django-money-rates==0.3.1
django-taggit==0.17.1
gunicorn==19.3.0
ipython==4.0.0
ipython-genutils==0.1.0
path.py==8.1
pexpect==3.3
pickleshare==0.5
psycopg2==2.6.1
py-moneyed==0.6.0
pycountry==1.15
simplegeneric==0.8.1
six==1.9.0
traitlets==4.0.0
wheel==0.24.0
